pyproject.toml
[tool.poetry]
name = "carsale-api"
version = "0.1.0"
description = ""
authors = ["Makima3 <natalia.kolchuk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.9"
django-dotenv = "^1.4.2"
django-countries = "^7.6.1"
django-filter = "^24.3"
pillow = "^11.0.0"
djangorestframework-simplejwt = "^5.3.1"
better-profanity = "^0.7.0"
drf-yasg = "1.21.8"
celery = "^5.4.0"
redis = "^5.2.0"
django-celery-results = "^2.5.1"
django-celery-beat = "^2.7.0"
requests = "^2.32.3"
djangochannelsrestframework = "^1.2.0"
channels-redis = "^4.2.1"
daphne = "^4.1.2"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
